Move spaces into respective directories
