+++ garage/tf/policies/categorical_conv_policy.py
@@ -2,7 +2,6 @@ import tensorflow as tf
 
 from garage.core import Serializable
 from garage.misc.overrides import overrides
-from garage.spaces.tf import Discrete
 from garage.tf.core import ConvNetwork
 from garage.tf.core import LayersPowered
 import garage.tf.core.layers as L
@@ -10,6 +9,7 @@ from garage.tf.distributions import Categorical
 from garage.tf.misc import tensor_utils
 from garage.tf.misc.tensor_utils import enclosing_scope
 from garage.tf.policies import StochasticPolicy
+from garage.tf.spaces import Discrete
 
 
 class CategoricalConvPolicy(StochasticPolicy, LayersPowered, Serializable):
