+++ garage/tf/policies/gaussian_mlp_policy.py
@@ -4,7 +4,6 @@ import tensorflow as tf
 from garage.core import Serializable
 from garage.misc import logger
 from garage.misc.overrides import overrides
-from garage.spaces.tf import Box as TfBox
 from garage.tf.core import LayersPowered
 from garage.tf.core import MLP
 import garage.tf.core.layers as L
@@ -12,6 +11,7 @@ from garage.tf.distributions import DiagonalGaussian
 from garage.tf.misc import tensor_utils
 from garage.tf.misc.tensor_utils import enclosing_scope
 from garage.tf.policies import StochasticPolicy
+from garage.tf.spaces import Box
 
 
 class GaussianMLPPolicy(StochasticPolicy, LayersPowered, Serializable):
@@ -56,7 +56,7 @@ class GaussianMLPPolicy(StochasticPolicy, LayersPowered, Serializable):
         :return:
         """
         Serializable.quick_init(self, locals())
-        assert isinstance(env_spec.action_space, TfBox)
+        assert isinstance(env_spec.action_space, Box)
         self.name = name
 
         with tf.variable_scope(name):
