+++ garage/tf/spaces/__init__.py
@@ -1,0 +1,7 @@
+"""TensorFlow extension of garage.spaces."""
+
+from garage.tf.spaces.box import Box
+from garage.tf.spaces.discrete import Discrete
+from garage.tf.spaces.product import Product
+
+__all__ = ["Box", "Discrete", "Product"]
