+++ garage/theano/spaces/product.py
